+++ contracts/interfaces/curations/IBardsCurationBase.sol
@@ -43,27 +43,27 @@ interface IBardsCurationBase {
      *
      * @param tokenId The token ID of the NFT to query the curation BPS for.
      *
-     * @return uint16[] The fee that is sent to the sellers.
+     * @return uint32[] The fee that is sent to the sellers.
      */
-    function sellerBpsesOf(uint256 tokenId) external view returns (uint16[] memory);
+    function sellerBpsesOf(uint256 tokenId) external view returns (uint32[] memory);
 
 	/**
      * @notice Returns the curation BPS associated with a given NFT, stored only once upon initial mint.
      *
      * @param tokenId The token ID of the NFT to query the curation BPS for.
      *
-     * @return uint16 curation BPS
+     * @return uint32 curation BPS
      */
-    function curationBpsOf(uint256 tokenId) external view returns (uint16);
+    function curationBpsOf(uint256 tokenId) external view returns (uint32);
 
 	/**
      * @notice Returns the staking BPS associated with a given NFT, stored only once upon initial mint.
      *
      * @param tokenId The token ID of the NFT to query thestaking BPS for.
      *
-     * @return uint16 staking BPS
+     * @return uint32 staking BPS
      */
-    function stakingBpsOf(uint256 tokenId) external view returns (uint16);
+    function stakingBpsOf(uint256 tokenId) external view returns (uint32);
 
 	/**
      * @notice Returns the curation data associated with a given NFT. This allows fetching the curation BPS and 
@@ -116,7 +116,7 @@ interface IBardsCurationBase {
      * @param tokenId The token Id of the NFT to set fee params.
      * @param sellerBpses The fee that is sent to the sellers.
      */
-    function setSellerBpsesParams(uint256 tokenId, uint16[] calldata sellerBpses) external;
+    function setSellerBpsesParams(uint256 tokenId, uint32[] calldata sellerBpses) external;
 
 	/**
      * @notice Sets fee parameters for a NFT
@@ -125,7 +125,7 @@ interface IBardsCurationBase {
      * @param curationBps The bps of curation
      * @param stakingBps The bps of staking
      */
-    function setBpsParams(uint256 tokenId, uint16 curationBps, uint16 stakingBps) external;
+    function setBpsParams(uint256 tokenId, uint32 curationBps, uint32 stakingBps) external;
 
     /**
      * @notice Sets curation fee parameters for a NFT
@@ -133,7 +133,7 @@ interface IBardsCurationBase {
      * @param tokenId The token Id of the NFT to set fee params.
      * @param curationBps The bps of curation
      */
-    function setCurationBpsParams(uint256 tokenId, uint16 curationBps) external;
+    function setCurationBpsParams(uint256 tokenId, uint32 curationBps) external;
 
     /**
      * @notice Sets staking fee parameters for a NFT
@@ -141,5 +141,5 @@ interface IBardsCurationBase {
      * @param tokenId The token Id of the NFT to set fee params.
      * @param stakingBps The bps of staking
      */
-    function setStakingBpsParams(uint256 tokenId, uint16 stakingBps) external;
+    function setStakingBpsParams(uint256 tokenId, uint32 stakingBps) external;
 }
